+++ oceanics-io-www/src/components/Oceanside.tsx
@@ -1,17 +1,17 @@
-/**
- * React and friends.
- */
-import React, { useEffect } from "react";
+import React from "react";
 import styled from "styled-components";
 
 /**
- * Hooks for data handling.
+ * Hooks for render loop and interaction.
  */
-import useOceansideWorld from "../hooks/useOceansideWorld";
-import useOceansideBoard from "../hooks/useOceansideBoard";
+import useOceansideWorld from "../hooks/useOceanside";
 import useWasmRuntime from "../hooks/useWasmRuntime";
-import useWorker from "../hooks/useWorker";
-import type {IWorld} from "../hooks/useOceansideWorld";
+import type { IWorld } from "../hooks/useOceanside";
+
+/**
+ * Class names for ref in CSS-in-JSS.
+ */
+const [WORLD, BOARD] = ["world", "board"];
 
 /**
  * Page-specific typings.
@@ -25,49 +25,24 @@ export interface ApplicationType extends IWorld {
 };
 
 /**
- * Webpack needs a static path at build time to make loadable chunks
- * from the worker script. There is probably a more clever way to do
- * this. 
- */
-const createBathysphereWorker = () => {
-  return new Worker(
-      new URL("../workers/oceanside.worker.ts", import.meta.url)
-  );
-}
-
-/**
  * Page component rendered by NextJS.
  */
 const Oceanside = ({ className, ...props }: ApplicationType) => {
   /**
-   * Single runtime for Oceanside context
+   * Main thread web assembly runtime.
    */
   const {runtime} = useWasmRuntime();
 
   /**
-   * Dedicated worker for performing numerical computation and text
-   * analysis in the background. 
+   * Synthetic terrain with digital elevation map and 
+   * probability table of feature types for world-building.
    */
-  const worker = useWorker("bathysphere", createBathysphereWorker());
-
-  /**
-   * Digital elevation map of the synthetic terrain, with a 
-   * probability table of feature types for world-building
-   */
-  const world = useOceansideWorld({...props, runtime, worker: worker.ref});
-  const board = useOceansideBoard({runtime, world})
-
-  /**
-   * Required to start the background WASM runtime. 
-   */
-  useEffect(() => {
-    worker.start(createBathysphereWorker());
-  }, []);
+  const {world, board} = useOceansideWorld({...props, runtime});
 
   return (
     <div className={className}>
-      <canvas {...world.canvas} className={"world"}/>
-      <canvas {...board.canvas} className={"board"}/>
+      <canvas {...world.canvas} className={WORLD}/>
+      <canvas {...board.canvas} className={BOARD}/>
     </div>
   );
 };
@@ -84,12 +59,12 @@ const StyledViewport = styled(Oceanside)`
   & canvas {
     image-rendering: crisp-edges;
   }
-  & .world {
+  & .${WORLD} {
     display: none;
     width: 256px;
     height: 256px;
   }
-  & .board {
+  & .${BOARD} {
     width: 100%;
     height: 512px;
   }
