+++ oceanics-io-www/src/hooks/useOceanside.tsx
@@ -1,14 +1,19 @@
 import { useEffect, useState, useRef, useCallback, useMemo } from "react";
-import type { MutableRefObject } from "react";
+import type { MouseEvent } from "react";
 
 import { eventCoordinates } from "../utils";
 import type { EventLocation, ModuleType } from "../utils";
-import type { MiniMap } from "oceanics-io-www-wasm";
+
+import useWorker from "../hooks/useWorker";
 
 /**
- * Messaging passing type
+ * Types generated by WasmBindGen
  */
+import type { MiniMap } from "oceanics-io-www-wasm";
+export type MapType = MiniMap;
+
 const ACTION = "parseIconSet";
+const DEFAULT_BLENDING = "#000000FF";
 
 export interface IWorldType {
   /**
@@ -36,6 +41,10 @@ export interface IWorldType {
    * Other wet tiles become mud depending on the tidal cycle and their elevation.
    */
   datum: number;
+  /**
+   * color of animation loop blending
+   */
+  backgroundColor?: string;
 }
 
 export interface IWorld extends IWorldType {
@@ -43,7 +52,6 @@ export interface IWorld extends IWorldType {
    * WASM package, already initialized
    */
   runtime: ModuleType | null;
-  worker: MutableRefObject<Worker | null>;
   icons: {
     sources: any;
     templates: any;
@@ -63,18 +71,24 @@ export interface IWorld extends IWorldType {
  * Tile asset references are used to pre-load all of the
  * sprite data for animations.
  */
-export const useOceansideWorld = ({
+export const useOceanside = ({
   grid,
   size,
   datum,
   runtime,
-  worker,
   icons,
+  backgroundColor = DEFAULT_BLENDING,
 }: IWorld) => {
   /**
    * Ref for clickable minimap that allows world navigation
    */
   const ref = useRef<HTMLCanvasElement | null>(null);
+
+  /**
+   * Dedicated background worker for number crunching and text
+   * analysis.
+   */
+  const worker = useWorker("oceanside", "../workers/oceanside.worker.ts");
 
   /**
    * When the runtime loads, create a pixel map
@@ -94,8 +108,8 @@ export const useOceansideWorld = ({
     if (!runtime || !ref.current) return null;
     const ctx = ref.current.getContext("2d");
     if (!ctx) {
-      console.error("Rendering Context is Null")
-      return null
+      console.error("Rendering Context is Null");
+      return null;
     }
     const offset = (size - grid.size) / 2;
     return new runtime.MiniMap(offset, offset / 2, size, datum, ctx, grid.size);
@@ -143,25 +157,25 @@ export const useOceansideWorld = ({
   );
 
   /**
-   * When we have a worker ready, we can start sending 
-   * messages to parse icon data and do some calculations 
+   * When we have a worker ready, we can start sending
+   * messages to parse icon data and do some calculations
    * of, for example, the feature probability distribution.
-   * 
+   *
    * We also need the map instance to exist, so that we can
    * insert the feature into the visualization.
    */
   useEffect(() => {
-    if (!map || !worker.current) return;
-    worker.current.addEventListener("message", listener, { passive: true });
-    worker.current.postMessage({
+    if (!map || !worker.ref.current) return;
+    worker.ref.current.addEventListener("message", listener, { passive: true });
+    worker.ref.current.postMessage({
       type: ACTION,
       data: [icons.sources, icons.templates, size],
     });
     return () => {
-      worker.current?.removeEventListener("message", listener);
-      worker.current?.terminate();
+      worker.ref.current?.removeEventListener("message", listener);
+      worker.ref.current?.terminate();
     };
-  }, [worker.current, map]);
+  }, [worker.ref.current, map]);
 
   /**
    * Re-populate the currently visible tiles when a mouse event fires
@@ -185,19 +199,89 @@ export const useOceansideWorld = ({
     [map]
   );
 
+  /**
+   * Ref for isometric view render target
+   */
+  const board = useRef<HTMLCanvasElement | null>(null);
+
+  /**
+   * Draw the visible area to the board canvas using the
+   * tile set object.
+   */
+  useEffect(() => {
+    if (!board.current || !tiles.length || !runtime || !map) return;
+
+    [board.current.width, board.current.height] = ["width", "height"]
+      .map((dim: string) =>
+        getComputedStyle(board.current as HTMLCanvasElement)
+          .getPropertyValue(dim)
+          .slice(0, -2)
+      )
+      .map((x: string) => parseInt(x) * window.devicePixelRatio);
+
+    const ctx = board.current.getContext(`2d`);
+    if (!ctx) return;
+
+    let requestId: number | null = null;
+
+    ctx.imageSmoothingEnabled = false; // disable interpolation
+
+    (function render() {
+      const { width, height } = board.current;
+      runtime.clear_rect_blending(ctx, width, height, backgroundColor);
+
+      if (!tiles.length)
+        throw TypeError(`Length of tile coordinate map is zero.`);
+      tiles.forEach((diagonal: number[], ii: number) => {
+        if (!diagonal.length)
+          throw TypeError(`Length of diagonal ${ii} is zero.`);
+        diagonal.forEach((tile, jj) => {
+          (map as MapType).drawTile(
+            ctx,
+            ii,
+            jj,
+            diagonal.length,
+            performance.now(),
+            width,
+            tile
+          );
+        });
+      });
+      requestId = requestAnimationFrame(render);
+    })();
+
+    return () => {
+      if (requestId) cancelAnimationFrame(requestId);
+    };
+  }, [map, tiles, runtime]);
+
   return {
-    canvas: {
-      width: size,
-      height: size,
-      ref,
-      onClick,
+    world: {
+      canvas: {
+        width: size,
+        height: size,
+        ref,
+        onClick,
+      },
+      map,
+      grid: {
+        ...grid,
+        tiles,
+      },
     },
-    map,
-    grid: {
-      ...grid,
-      tiles,
+    board: {
+      canvas: {
+        ref: board,
+        onClick: (event: MouseEvent) => {
+          try {
+            console.log({ event });
+          } catch (err) {
+            console.error(err);
+          }
+        },
+      },
     },
   };
 };
 
-export default useOceansideWorld;
+export default useOceanside;
