+++ oceanics-io-www/src/workers/oceanside.worker.ts
@@ -43,8 +43,6 @@ const parseIconSet = (
   nodes.map(({ slug }) => [slug, slug])
 )));
 
-
-
 /**
  * Event handler
  */
